Patched for Chrome 87
Note that this was only briefly tested to a limited extent on OS X 10.13 with Chrome 87.0.4280.67; I did not test this at all on Windows, although I would imagine that the same selectors were used there.

The new height parameter in the body style is not correctly terminated (missing semicolon). I think that may change in the future, and if/when it does, then this will probably break and require another (minor) update to fix it again.